Split saucectl-exporter
